+++ moshi-ir/moshi-compiler-plugin/testData/diagnostic/InnerClassesNotSupported.kt
@@ -1,6 +1,4 @@
-// RUN_PIPELINE_TILL: FIR2IR
-// DISABLE_GENERATED_FIR_TAGS
-// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+// RENDER_DIAGNOSTICS_FULL_TEXT
 
 package test
 
@@ -10,4 +8,3 @@ class Outer {
   <!MOSHI_ERROR!>@JsonClass(generateAdapter = true)
   inner class InnerClass(val a: Int)<!>
 }
-