+++ moshi-ir/moshi-compiler-plugin/testData/diagnostic/InterfacesNotSupported.kt
@@ -1,6 +1,4 @@
-// RUN_PIPELINE_TILL: FIR2IR
-// DISABLE_GENERATED_FIR_TAGS
-// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+// RENDER_DIAGNOSTICS_FULL_TEXT
 
 package test
 
@@ -8,4 +6,3 @@ import com.squareup.moshi.JsonClass
 
 <!MOSHI_ERROR!>@JsonClass(generateAdapter = true)
 interface Interface<!>
-