+++ moshi-ir/moshi-compiler-plugin/testData/diagnostic/InvalidConstructorParam.kt
@@ -1,6 +1,4 @@
-// RUN_PIPELINE_TILL: FIR2IR
-// DISABLE_GENERATED_FIR_TAGS
-// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+// RENDER_DIAGNOSTICS_FULL_TEXT
 
 package test
 
@@ -15,4 +13,3 @@ sealed class BaseType {
   @TypeLabel("a")
   class TypeA : BaseType()
 }<!>
-