+++ moshi-ir/moshi-compiler-plugin/testData/diagnostic/InvisibleFallbackAdapterConstructor.kt
@@ -1,6 +1,4 @@
-// RUN_PIPELINE_TILL: FIR2IR
-// DISABLE_GENERATED_FIR_TAGS
-// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+// RENDER_DIAGNOSTICS_FULL_TEXT
 
 package test
 
@@ -12,15 +10,15 @@ import dev.zacsweers.moshix.sealed.annotations.DefaultObject
 import dev.zacsweers.moshix.sealed.annotations.FallbackJsonAdapter
 import dev.zacsweers.moshix.sealed.annotations.TypeLabel
 
-<!MOSHI_ERROR!>class BaseTypeFallback private constructor() : JsonAdapter<String>() {
+class BaseTypeFallback private constructor() : JsonAdapter<String>() {
   override fun fromJson(reader: JsonReader): String? {
     return null
   }
 
   override fun toJson(writer: JsonWriter, value: String?) {}
-}<!>
+}
 
-@FallbackJsonAdapter(BaseTypeFallback::class)
+<!MOSHI_ERROR!>@FallbackJsonAdapter(BaseTypeFallback::class)
 @JsonClass(generateAdapter = true, generator = "sealed:type")
 sealed class BaseType {
   @TypeLabel("a")
@@ -28,5 +26,4 @@ sealed class BaseType {
 
   @DefaultObject
   object TypeB : BaseType()
-}
-
+}<!>