+++ moshi-ir/moshi-compiler-plugin/testData/diagnostic/NonRuntimeQualifier.kt
@@ -1,6 +1,4 @@
-// RUN_PIPELINE_TILL: FIR2IR
-// DISABLE_GENERATED_FIR_TAGS
-// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+// RENDER_DIAGNOSTICS_FULL_TEXT
 
 package test
 
@@ -19,4 +17,3 @@ annotation class UpperCase
 
 @JsonClass(generateAdapter = true)
 class ClassWithQualifier(<!MOSHI_ERROR!>@UpperCase<!> val a: Int)
-