+++ moshi-ir/moshi-compiler-plugin/testData/diagnostic/PrivateProperties.kt
@@ -1,6 +1,4 @@
-// RUN_PIPELINE_TILL: FIR2IR
-// DISABLE_GENERATED_FIR_TAGS
-// RENDER_IR_DIAGNOSTICS_FULL_TEXT
+// RENDER_DIAGNOSTICS_FULL_TEXT
 
 package test
 
@@ -11,4 +9,3 @@ class PrivateProperties {
   private var a: Int = -1
   private var b: Int = -1
 }<!>
-